post

+++ app/Http/Controllers/TasksController.php
@@ -18,11 +18,22 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $task = Task::paginate(10);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('Tasks.index', [
-            'tasks' => $task,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('Tasks.index', $data);
+        }else {
+            return view('welcome');
+            }
+
+        
     }
 
     /**
@@ -57,6 +68,10 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+        ]);
 
         return redirect('/');
     }
@@ -71,9 +86,16 @@ class TasksController extends Controller
     {
          $task = Task::find($id);
 
-        return view('Tasks.show', [
-            'task' => $task,
-        ]);
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+        $data = [
+            'user' => $user,
+            'tasks' => $tasks,
+        ];
+
+        $data += $this->counts($user);
+
+        return view('Tasks.show', $data);
     }
 
     /**
@@ -122,8 +144,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
          $task = Task::find($id);
-        $task->delete();
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
-        return redirect('/');
+        return redirect()->back();
     }
 }
